requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.16
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.2
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.8.1
fastapi==0.115.4
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.2.10
langchain-community==0.2.10
langchain-core==0.2.24
langchain-text-splitters==0.2.2
langsmith==0.1.137
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-langchain==0.2.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-readers-wikipedia==0.2.0
llama-index-vector-stores-chroma==0.2.1
llama-parse==0.5.12
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==2.5.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.52.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
posthog==3.7.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.12
pyparsing==3.2.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.41.2
streamlit==1.39.0
streamlit-aggrid==1.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.1
trulens-apps-langchain==1.2.1
trulens-apps-llamaindex==1.2.1
trulens-core==1.1.0
trulens-dashboard==1.2.1
trulens-feedback==1.2.1
trulens-providers-openai==1.2.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
validators==0.34.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
wikipedia==1.4.0
Wikipedia-API==0.7.1
wordcloud==1.9.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.0
zipp==3.20.2
